requirements.txt
requests
html2text
airtable-python-wrapper
python-dotenv